Created exchanges unit tests

+++ src/main/java/org/cryptodata/service/exchange/asset/models/AssetData.java
@@ -2,9 +2,11 @@ package org.cryptodata.service.exchange.asset.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigInteger;
+
 public record AssetData(
         @JsonProperty("wallet_address") String walletAddress,
-        Integer balance,
+        BigInteger balance,
         AssetPlatform platform,
         AssetCurrency currency
 ) {
